+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Release 1.6
 
 - Upgraded pre-commit hooks (pre-commit-hooks to `v5.0.0` and ruff-pre-commit to `v0.6.0`)
+- Added [run-test.sh](run-tests.sh) script that runs all checks on code
 - Added support for Python 3.12 and 3.13 by upgrading Pylint and disabling/fixing Pylint errors
 - Corrected and improved language consistency in [readme](README.md) and `CHANGELOG.md`